refactor user, location app(total backend branch coverage 97%)
